feat: Add MongoDB connection logic
This commit adds the logic to establish a connection to MongoDB using the provided MONGODB_URI environment variable. If a connection is already established, it will reuse the existing connection. Otherwise, it will attempt to connect to the MongoDB server. If the connection is successful, a message will be logged indicating the successful connection. If there is an error connecting to MongoDB, an error message will be logged and the process will exit with an error code.